requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cloudscraper==1.2.58
docopt==0.6.2
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysqlclient==2.0.3
nimporter==1.0.4
orjson==3.6.4
Pillow==8.4.0
protobuf==3.19.0
pycodestyle==2.8.0
pyparsing==3.0.0
python-dotenv==0.19.1
requests==2.26.0
requests-toolbelt==0.9.1
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
yarg==0.1.9
